requirements.txt
streamlit>=1.22.0
pillow>=11.2.0
opencv-python-headless>=4.11.0
numpy>=1.26.0 